Scriptable Objects
Refactored CreateWeapon to CreateGun and CreateMelee, edited Gun and
Melee accordingly, and created a test gun and melee asset.
